Replace workflow execution service with event service
